save return int64 instead

+++ src/model/Model.h
@@ -177,7 +177,7 @@ public:
     * in the record list (Data_Set) to the database.
     */
     template<class DATA>
-    int save(std::vector<DATA>& rows)
+    int64 save(std::vector<DATA>& rows)
     {
         this->Savepoint();
         for (auto& r : rows)
@@ -192,7 +192,7 @@ public:
     }
 
     template<class DATA>
-    int save(std::vector<DATA*>& rows)
+    int64 save(std::vector<DATA*>& rows)
     {
         this->Savepoint();
         for (auto& r : rows) this->save(r);
